# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Haife API",
    "version": "1.0.0",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/page": {
      "post": {
        "tags": [
          "Page"
        ],
        "summary": "新增頁面",
        "description": "傳入頁面資訊並建立頁面",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "description": "使用者權杖",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "頁面新增資訊",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "頁面名稱"
                },
                "displayName": {
                  "type": "string",
                  "example": "頁面顯示名稱"
                },
                "path": {
                  "type": "string",
                  "example": "路徑"
                },
                "parentId": {
                  "type": "string",
                  "example": "父頁面編號"
                },
                "sort": {
                  "type": "string",
                  "example": "排序"
                },
                "icon": {
                  "type": "string",
                  "example": "圖標"
                }
              },
              "required": [
                "name",
                "displayName",
                "path",
                "parentId",
                "sort",
                "icon"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "新增成功"
          },
          "400": {
            "description": "新增失敗"
          },
          "403": {
            "description": "使用者沒有角色頁面權限"
          }
        }
      },
      "get": {
        "tags": [
          "Page"
        ],
        "summary": "頁面查詢 (含分頁與搜尋)",
        "description": "列出所有頁面資訊 (含分頁與搜尋)",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "description": "使用者權杖",
            "type": "string"
          },
          {
            "name": "rowsPerPage",
            "in": "query",
            "required": false,
            "description": "每頁筆數",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "目前頁數",
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "description": "排序欄位",
            "type": "string"
          },
          {
            "name": "sortType",
            "in": "query",
            "required": false,
            "description": "排序方式",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "description": "搜尋資訊 (可空值)",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "查詢成功"
          },
          "400": {
            "description": "查詢失敗"
          },
          "403": {
            "description": "使用者沒有角色頁面權限"
          }
        }
      }
    },
    "/api/page/{uuid}": {
      "patch": {
        "tags": [
          "Page"
        ],
        "summary": "修改頁面",
        "description": "根據頁面編號，傳入頁面資訊並修改頁面",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "頁面編號"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "description": "使用者權杖",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "頁面修改資訊",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "頁面名稱"
                },
                "displayName": {
                  "type": "string",
                  "example": "頁面顯示名稱"
                },
                "path": {
                  "type": "string",
                  "example": "路徑"
                },
                "parentId": {
                  "type": "string",
                  "example": "父頁面編號"
                },
                "sort": {
                  "type": "string",
                  "example": "排序"
                },
                "icon": {
                  "type": "string",
                  "example": "圖標"
                }
              },
              "required": [
                "name",
                "displayName",
                "path",
                "parentId",
                "sort",
                "icon"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "修改成功"
          },
          "400": {
            "description": "修改失敗"
          },
          "403": {
            "description": "使用者沒有角色頁面權限"
          }
        }
      },
      "delete": {
        "tags": [
          "Page"
        ],
        "summary": "刪除頁面",
        "description": "根據頁面編號刪除頁面",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "頁面編號"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "description": "使用者權杖",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "刪除成功"
          },
          "400": {
            "description": "刪除失敗"
          },
          "403": {
            "description": "使用者沒有角色頁面權限"
          }
        }
      }
    },
    "/api/page/parent": {
      "get": {
        "tags": [
          "Page"
        ],
        "summary": "查詢所有父頁面",
        "description": "列出所有父頁面資訊",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "description": "使用者權杖",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "查詢成功"
          },
          "400": {
            "description": "查詢失敗"
          },
          "403": {
            "description": "使用者沒有角色頁面權限"
          }
        }
      }
    },
...
  }
}